Set http.route attribute as path template
We have seen issues with the current spans exported from the http
library in observability backends, due to too granular span names when
the routes contain path variables.

As per the opentelemetry specification, this PR extracts the path
template from the UriRouteMatch object, and sets it as the
`http.route` attribute.

https://opentelemetry.io/docs/reference/specification/trace/semantic_conventions/http/#http-server-semantic-conventions