Upgrade Open Telemetry to 1.36 (#528)
* Upgrade Open Telemetry to 1.36
* Document breaking changes and deprecate HttpClientSender

The main Open Telemetry BOM dependency is updated to 1.3.6. This has a
cascading effect as it requires updating the Open Telemetry
instrumentation libraries to 1.31, and the Zipkin dependencies to 3.x.
Minor changes are implemented in the codebase to work with these
dependency updates.

Since the required Zipkin update is to a new major version, the module
version is bumped to 7.0.0.