fix: Remove core processor from the mn opentelemetry http library
Closes #553

The Micronaut OpenTelemetry Http plugin was exposing an API dependency on the processor classes.
This in turn pulled in 7Mb of unrequired dependencies.

This PR fixes this issue, and does a tiny bit of cleanup.

In general, this module seems to have an excessive use of `api` for dependencies.
But I don't think we can do much more without causing downstream breakages.

I ran the guide mentioned in the issue, and then re-ran it with a snapshot version of this PR, the following Build Scans were the result:

`bookcatalog`
- before https://ge.micronaut.io/s/ofkuix744eine
- after https://ge.micronaut.io/s/s6wap7cgohfzs

`bookinventory`
- before https://ge.micronaut.io/s/vokg7vtiyhztg
- after https://ge.micronaut.io/s/67p35n6pkovce

`bookrecommendation`
- before https://ge.micronaut.io/s/6ozkmgcfwtgb4
- after https://ge.micronaut.io/s/i3hi5u2iwgjzs